PhoneSessionManager + PhoneMessageReceiver 통합
delegate 충돌 해결
PhoneMessageReceiver 주석처리
