+++ umm/umm/ConversationMain/PhoneSessionManager.swift
@@ -15,18 +15,21 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
     
     @Published var startTime: String = ""
     @Published var exitTime: String = ""
+    @Published var receivedSuggestions: [WordSuggestion] = []
+    @Published var conversationSessions: [ConversationSession] = []
     
     private let speechRecognizer = SFSpeechRecognizer(locale: Locale(identifier: "ko-KR"))
     private var recognitionTask: SFSpeechRecognitionTask?
     
-    // ✅ PhoneMessageReceiver 인스턴스를 내부에서 직접 생성
-    private let messageReceiver = PhoneMessageReceiver()
+    // 현재 진행 중인 세션
+    private var currentSession: ConversationSession?
     
     private override init() {
         super.init()
         setupSpeechRecognition()
         activateSession()
         registerBackgroundTask()
+        loadSavedSessions()
     }
     
     private func setupSpeechRecognition() {
@@ -45,20 +48,38 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
     private func activateSession() {
         if WCSession.isSupported() {
             let session = WCSession.default
-            session.delegate = self
+            session.delegate = self  // ✅ 하나의 delegate로 통합
             session.activate()
         }
     }
     
+    // 백그라운드 작업 등록
     private func registerBackgroundTask() {
         BGTaskScheduler.shared.register(forTaskWithIdentifier: "com.yourapp.audioprocessing", using: nil) { task in
             self.handleBackgroundAudioProcessing(task: task as! BGProcessingTask)
         }
     }
     
-    // ✅ 메시지를 받았을 때 messageReceiver로 전달해줘야 시간 데이터 저장됨
-    func session(_ session: WCSession, didReceiveMessage message: [String: Any], replyHandler: @escaping ([String: Any]) -> Void) {
+    // MARK: - WCSessionDelegate 구현
+    
+    func session(_ session: WCSession, activationDidCompleteWith activationState: WCSessionActivationState, error: Error?) {
+        print("✅ WCSession 활성화 완료. 상태: \(activationState.rawValue)")
+    }
+    
+    func sessionDidBecomeInactive(_ session: WCSession) {
+        print("🔄 세션 비활성화됨")
+    }
+    
+    func sessionDidDeactivate(_ session: WCSession) {
+        print("🛑 세션 종료됨. 새로운 세션 활성화 가능")
+        WCSession.default.activate()
+    }
+    
+    // ✅ 통합된 메시지 수신 처리 (Reply Handler 있는 버전)
+    func session(_ session: WCSession, didReceiveMessage message: [String : Any], replyHandler: @escaping ([String : Any]) -> Void) {
         print("📨 [Manager] 메시지 수신: \(message)")
+        
+        // 오디오 데이터 즉시 처리
         if let audioData = message["audioData"] as? Data,
            message["needsImmediateProcessing"] as? Bool == true {
             print("📥 즉시 처리용 오디오 데이터 수신됨")
@@ -71,43 +92,60 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
             return
         }
         
-        // ✅ 시간, 제안 메시지 등은 messageReceiver로 전달
-        messageReceiver.session(session, didReceiveMessage: message)
-        
-        // ✅ (선택사항) startTime, exitTime을 UI에서 바로 쓸 수 있도록 별도로 저장
+        // 시간 및 제안 메시지 처리
         DispatchQueue.main.async {
-            if let startTime = message["startTime"] as? String {
-                self.startTime = startTime
-                print("✅ 받은 startTime: \(startTime)")
-            } else if let exitTime = message["exitTime"] as? String {
-                self.exitTime = exitTime
-                print("✅ 받은 exitTime: \(exitTime)")
+            // 시작 시간 처리
+            if let startTimeString = message["startTime"] as? String {
+                self.startTime = startTimeString
+                self.handleStartTime(startTimeString)
+                print("✅ 받은 startTime: \(startTimeString)")
+            }
+            
+            // 종료 시간 처리
+            if let exitTimeString = message["exitTime"] as? String {
+                self.exitTime = exitTimeString
+                self.handleExitTime(exitTimeString)
+                print("✅ 받은 exitTime: \(exitTimeString)")
+            }
+            
+            // WordSuggestion 배열 처리
+            if let data = message["suggestions"] as? Data,
+               let decoded = try? JSONDecoder().decode([WordSuggestion].self, from: data) {
+                self.handleReceivedSuggestions(decoded)
             }
         }
         
-        // ✅ replyHandler는 한 번만!
         replyHandler([:])
     }
     
-    
+    // ✅ 통합된 메시지 수신 처리 (Reply Handler 없는 버전)
     func session(_ session: WCSession, didReceiveMessage message: [String: Any]) {
         print("📨 [Manager - NoReply] 메시지 수신: \(message)")
 
-        // ✅ 메시지를 messageReceiver로 전달
-        messageReceiver.session(session, didReceiveMessage: message)
-
-        // ✅ startTime, exitTime 업데이트
         DispatchQueue.main.async {
-            if let startTime = message["startTime"] as? String {
-                self.startTime = startTime
-                print("✅ 받은 startTime (noReply): \(startTime)")
-            } else if let exitTime = message["exitTime"] as? String {
-                self.exitTime = exitTime
-                print("✅ 받은 exitTime (noReply): \(exitTime)")
+            // 시작 시간 처리
+            if let startTimeString = message["startTime"] as? String {
+                self.startTime = startTimeString
+                self.handleStartTime(startTimeString)
+                print("✅ 받은 startTime (noReply): \(startTimeString)")
+            }
+            
+            // 종료 시간 처리
+            if let exitTimeString = message["exitTime"] as? String {
+                self.exitTime = exitTimeString
+                self.handleExitTime(exitTimeString)
+                print("✅ 받은 exitTime (noReply): \(exitTimeString)")
+            }
+            
+            // WordSuggestion 배열 처리
+            if let data = message["suggestions"] as? Data,
+               let decoded = try? JSONDecoder().decode([WordSuggestion].self, from: data) {
+                self.handleReceivedSuggestions(decoded)
             }
         }
     }
     
+    // ✅ 백그라운드 파일 수신 처리
     func session(_ session: WCSession, didReceive file: WCSessionFile) {
         print("📥 백그라운드 파일 수신됨: \(file.fileURL.lastPathComponent)")
         
@@ -137,16 +175,118 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
         }
     }
     
-    private func handleBackgroundAudioProcessing(task: BGProcessingTask) {
-        task.expirationHandler = {
-            task.setTaskCompleted(success: false)
+    // MARK: - 대화 세션 관리 (기존 PhoneMessageReceiver 기능 통합)
+    
+    // 시작 시간 처리 - 새로운 ConversationSession 시작
+    private func handleStartTime(_ startTimeString: String) {
+        let formatter = ISO8601DateFormatter()
+        guard let startTime = formatter.date(from: startTimeString) else {
+            print("❌ 시작 시간 파싱 실패: \(startTimeString)")
+            return
         }
         
-        processPendingAudioFiles { success in
-            task.setTaskCompleted(success: success)
+        // 새로운 세션 시작
+        currentSession = ConversationSession(
+            startTime: startTime,
+            endTime: nil,
+            groups: []
+        )
+        
+        print("✅ 새로운 대화 세션 시작: \(startTime)")
+    }
+    
+    // 종료 시간 처리 - 현재 세션 완료
+    private func handleExitTime(_ exitTimeString: String) {
+        let formatter = ISO8601DateFormatter()
+        guard let exitTime = formatter.date(from: exitTimeString),
+              var session = currentSession else {
+            print("❌ 종료 시간 처리 실패")
+            return
+        }
+        
+        // 현재 세션에 종료 시간 설정
+        session.endTime = exitTime
+        
+        // 완료된 세션을 배열에 추가 (최신순으로 정렬)
+        conversationSessions.insert(session, at: 0)
+        
+        // 현재 세션 초기화
+        currentSession = nil
+        
+        // 저장
+        saveSessions()
+        
+        print("✅ 대화 세션 완료: \(session.startTime) ~ \(exitTime)")
+        print("총 그룹 수: \(session.groups.count)")
+    }
+    
+    // WordSuggestion 배열을 받아서 현재 세션에 그룹으로 추가
+    private func handleReceivedSuggestions(_ suggestions: [WordSuggestion]) {
+        guard !suggestions.isEmpty else { return }
+        
+        // 현재 세션이 없으면 임시로 생성 (혹시 시작 메시지를 놓친 경우)
+        if currentSession == nil {
+            currentSession = ConversationSession(
+                startTime: Date(),
+                endTime: nil,
+                groups: []
+            )
+            print("⚠️ 시작 메시지 없이 제안 수신, 임시 세션 생성")
+        }
+        
+        // 첫 번째 suggestion의 meaning에서 한국어 키워드 추출
+        let keyword = extractKeywordFromSuggestions(suggestions)
+        
+        let newGroup = WordSuggestionGroup(
+            keyword: keyword,
+            suggestions: suggestions
+        )
+        
+        currentSession?.groups.append(newGroup)
+        
+        // UI 업데이트용
+        receivedSuggestions = suggestions
+        
+        print("📥 그룹 추가됨 - 키워드: \(keyword), 제안 수: \(suggestions.count)")
+    }
+    
+    // suggestion들로부터 키워드 추출하는 헬퍼 함수
+    private func extractKeywordFromSuggestions(_ suggestions: [WordSuggestion]) -> String {
+        if let firstSuggestion = suggestions.first {
+            let meaning = firstSuggestion.meaning
+            if let firstWord = meaning.components(separatedBy: ",").first?.trimmingCharacters(in: .whitespaces) {
+                return firstWord
+            }
+        }
+        return "알 수 없음"
+    }
+    
+    // MARK: - 세션 데이터 관리
+    
+    // 세션 저장
+    private func saveSessions() {
+        if let encoded = try? JSONEncoder().encode(conversationSessions) {
+            UserDefaults.standard.set(encoded, forKey: "ConversationSessions")
         }
     }
     
+    // 저장된 세션 불러오기
+    private func loadSavedSessions() {
+        if let data = UserDefaults.standard.data(forKey: "ConversationSessions"),
+           let decoded = try? JSONDecoder().decode([ConversationSession].self, from: data) {
+            conversationSessions = decoded
+        }
+    }
+    
+    // 세션 삭제
+    func deleteSession(at indexSet: IndexSet) {
+        conversationSessions.remove(atOffsets: indexSet)
+        saveSessions()
+    }
+    
+    // MARK: - 음성 인식 처리
+    
+    // ✅ 핵심 음성 인식 처리 로직
     private func processAudioData(_ audioData: Data, completion: @escaping (String) -> Void) {
         guard let speechRecognizer = speechRecognizer, speechRecognizer.isAvailable else {
             completion("음성인식을 사용할 수 없습니다")
@@ -182,9 +322,11 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
         }
     }
     
+    // MARK: - Watch 통신
+    
     func sendTextToWatch(_ text: String) {
         print("📡 Watch 연결 상태: \(WCSession.default.isReachable)")
-        
+
         if WCSession.default.isReachable {
             WCSession.default.sendMessage(["recognizedText": text], replyHandler: nil) { error in
                 print("❌ 텍스트 전송 실패: \(error.localizedDescription)")
@@ -192,7 +334,7 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
             }
             print("📤 텍스트 전송 완료: \(text)")
         } else {
-            print("⚠️ Watch에 연결되지 않음, Context로 전송합니다")
+            print("⚠️ Watch에 연결되지 않았습니다. Context로 전송합니다.")
             sendViaContext(text: text)
         }
     }
@@ -203,6 +345,18 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
             print("📤 Context로 전송됨: \(text)")
         } catch {
             print("❌ Context 전송 실패: \(error)")
+        }
+    }
+    
+    // MARK: - 백그라운드 처리
+    
+    private func handleBackgroundAudioProcessing(task: BGProcessingTask) {
+        task.expirationHandler = {
+            task.setTaskCompleted(success: false)
+        }
+        
+        processPendingAudioFiles { success in
+            task.setTaskCompleted(success: success)
         }
     }
     
@@ -247,10 +401,7 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
         }
     }
     
-    // ✅ 필요 시 외부에서 messageReceiver에 접근할 수 있도록 getter 제공
-    var messageReceiverInstance: PhoneMessageReceiver {
-        return messageReceiver
-    }
+    // MARK: - 기존 호환성 메서드
     
     func recognizeSpeech(from url: URL) {
         do {
@@ -261,20 +412,5 @@ class PhoneSessionManager: NSObject, WCSessionDelegate, ObservableObject {
         } catch {
             print("❌ 파일 읽기 실패: \(error)")
         }
-    }
-    
-    
-    // MARK: - 필수 WCSessionDelegate 구현
-    
-    func session(_ session: WCSession, activationDidCompleteWith activationState: WCSessionActivationState, error: Error?) {
-        print("✅ WCSession 활성화 완료. 상태: \(activationState.rawValue), 에러: \(String(describing: error))")
-    }
-    
-    func sessionDidBecomeInactive(_ session: WCSession) {
-        print("ℹ️ 세션 비활성화됨")
-    }
-    
-    func sessionDidDeactivate(_ session: WCSession) {
-        print("ℹ️ 세션 비활성화 해제됨 → 새로운 세션으로 활성화 가능")
     }
 }
